adds tests
