+++ backend/app/models.py
@@ -1,18 +1,16 @@
 from datetime import datetime
-from flask_sqlalchemy import SQLAlchemy
-
-db = SQLAlchemy()
+from .extensions import db  # Use relative import from extensions
 
 STATUS_VALUES = ("todo", "in_progress", "done")
 
 class Task(db.Model):
     __tablename__ = "tasks"
-
+    # ... your existing Task model code ...
     id = db.Column(db.Integer, primary_key=True)
     title = db.Column(db.String(120), nullable=False)
     description = db.Column(db.String(500))
     status = db.Column(db.String(20), default="todo", nullable=False)
-    due_date = db.Column(db.String(25))  # ISO date string (YYYY-MM-DD) or empty
+    due_date = db.Column(db.String(25))
     created_at = db.Column(db.DateTime, default=datetime.utcnow, nullable=False)
     updated_at = db.Column(db.DateTime, default=datetime.utcnow, onupdate=datetime.utcnow, nullable=False)
