+++ backend/requirements.txt
@@ -1,4 +1,6 @@
-flask==3.0.3
-flask-cors==4.0.1
-flask-sqlalchemy==3.1.1
-python-dotenv==1.0.1
+flask
+flask-cors
+flask-sqlalchemy
+python-dotenv
+pytest
+faker
