+++ backend/run.py
@@ -1,0 +1,7 @@
+import os
+from app import create_app
+
+if __name__ == "__main__":
+    app = create_app()
+    port = int(os.getenv("PORT", "5000"))
+    app.run(host="0.0.0.0", port=port, debug=True)
